requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
Flask-Limiter==2.4.0
python-dotenv==0.19.0
twilio==7.3.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
PyJWT==2.1.0
python-dateutil==2.8.2
sshtunnel==0.4.0
paramiko==2.11.0
PyMySQL==1.1.0 